+++ CRM/Contact/Page/DataProcessorContactSummaryTab.php
@@ -1,12 +1,10 @@
 <?php
-
-use CRM_Dataprocessor_ExtensionUtil as E;
-
 /**
- * Main page for Data Processor Output dashlet
- *
+ * @author Jaap Jansma <jaap.jansma@civicoop.org>
+ * @license AGPL-3.0
  */
-class CRM_DataprocessorDashlet_Page_Dashlet extends CRM_Core_Page {
+
+class CRM_Contact_Page_DataProcessorContactSummaryTab extends CRM_Core_Page {
 
   /**
    * @var int
@@ -35,11 +33,14 @@ class CRM_DataprocessorDashlet_Page_Dashlet extends CRM_Core_Page {
    */
 
   protected function preProcess() {
-    $this->dataProcessorName = CRM_Utils_Request::retrieve('data_processor', 'String');
+    $this->dataProcessorName = CRM_Utils_Request::retrieve('data_processor', 'String', $this, true);
+    $contact_id = CRM_Utils_Request::retrieve('contact_id', 'Integer', $this, true);
 
     $this->dataProcessor = civicrm_api3('DataProcessor', 'getsingle', array('name' => $this->dataProcessorName));
     $this->dataProcessorClass = CRM_Dataprocessor_BAO_DataProcessor::dataProcessorToClass($this->dataProcessor);
     $this->assign('dataProcessorName', $this->dataProcessorName);
+    $this->assign('contact_id', $contact_id);
+    $this->assign('url', CRM_Utils_System::url("civicrm/dataprocessor_contact_summary/{$this->dataProcessorName}", array('contact_id' => $contact_id, 'reset' => '1', 'snippet' => 'json')));
   }
 
   /**