+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 * Respect selected permissions for outputs 
 * Allow to specify "Is Empty" for various filters.
 * Allow to limit ContactFilter to only show contacts from specific groups.
-* it is now also possible to output a data processor on CiviCRMs dahsboard.
+* Output a data processor as a dashboard.
+* Output a data processor as a tab on the contact summary screen.
 * Added field outputs for simple calculations (substract and total).
 
 # Version 1.0.7